Cleaned up my code - check WhatsApp for detailedNotes
