+++ Code/utils.py
@@ -178,6 +178,7 @@ def loadData(month="March", pruned=False, combined = False):
     pathName+=".json"
 
     #Delete after usage.
+    #Currently using another json, so we can test the function, that was written.
     if month == "June":
         pathName = "../Data/"+month+"/trilatData.json"
 
@@ -703,840 +704,6 @@ def rewriteMarchData(month="March"):
     with open("../Data/"+month+"/associatedMarchData_2.json","w+") as f:
         json.dump(finalData, f)
 
-
-def LoadSeqData(month="June"):
-    '''
-        Idea is that if we give the machine the distances it is away from other nodes and the signal strenght, we would be able to predict the distance it is away from
-        the actual node.
-        So, basically trying to reproduce the 'proximity' filter, with a neural network -> thus we would get an even better distance result, that would count the
-        noise / vegetation in as well. Since, the distance ~ signalStr equation in itself is not enough to make the prediction to the actual distance.
-    '''
-    if(month =="June"):
-        # Load in the June JSON
-        pathName = "../Data/"+month+"/newData.json"
-        with open(pathName,"r") as f:
-            data = json.load(f)
-        nodes = loadNodes(rewriteUTM=True)
-
-    X = data["X"]
-    y = data["y"]
-    newX = []
-    newY = []
-    newZ = []
-    #Going to gather the new data pairs in the form of:
-    #Elements of newX will look like: [
-    #                                   [signalStrength of the current node in the batch],
-    #                                   [all relative node distances in the same batch, that also have relevant signal],
-    #                                   [all relative node signals in the same batch, that are relevant]
-    #                                   [batch_number]]
-    #Elements of newY will look like: [distance between current node and the drone]
-    n = 0
-    for index, item in enumerate(data["X"]):
-
-        #Keys
-        keys = item["data"].keys()
-        #Get the drone position in lat,long
-        dronePos = [data["y"][index][0],data["y"][index][1]]
-        #Convert those positions to utm coordinates
-        utmDronePos = utm.from_latlon(dronePos[0],dronePos[1])
-        utmDronePosArr = np.array([np.float64(utmDronePos[0]),np.float64(utmDronePos[1])])
-
-        #getNodedistances returns two lists. One of them contains [[relative node distances in the same batch],[position of current node], [signal of current node]]
-        #The skipKeys will have all the NodeIDs of the nodes that need to be skipped because they are
-        #outside of the range for the signals
-        replace, skipKeys = getNodedistances(item["data"])
-
-        #Gets all the signals in the batch
-        signals = getSignals(item["data"])
-
-        for key in keys:
-            if(key == "3288000000"): key="3288e6"
-            #If the key has wrong outside of range signal, we skip it
-            if key in skipKeys:continue
-
-            #If the drone position is not recorded well, we skip it
-            if(dronePos == (0,0) or dronePos == (0.754225181062586, -12.295563892977972) or dronePos == (4.22356791831445, -68.85519277364834)): continue
-
-            #We gather all the signals that are in the batch for the current node
-            #and skip the outside of the range signals
-            sSignals = []
-            for newKey in keys:
-                if(newKey == "3288000000"): key="3288e6"
-                if newKey == key: continue
-                if newKey in skipKeys:continue
-                sSignals.append(signals[newKey])
-            #Need at least 3 nodes, so we take out everything else - lost about 48 data points from the 4600.
-            if len(sSignals) < 2: continue
-            #current node position in latlon, then convert it into utm, then get the distance between them
-            sNodePos = item["data"][key][1]
-            utmsNodePos = utm.from_latlon(sNodePos[0],sNodePos[1])
-            utmsNodePosArr = np.array([np.float64(utmsNodePos[0]),np.float64(utmsNodePos[1])])
-            actualDist = np.linalg.norm(utmsNodePosArr-utmDronePosArr)
-
-            #Eventually add these data pairs into the newX and newY, which will be saved for later training and modelling.
-            newX.append([item['data'][key][2],item['data'][key][0], sSignals, index, key])
-            newY.append(actualDist)
-            newZ.append(dronePos)
-    assert(len(newX) == len(newY))
-    finalData = {}
-    finalData['X']=newX
-    finalData['y']=newY
-    finalData['Z']=newZ
-
-    #Save the file in a json, that will be used for modelling / training
-    with open("../Data/"+month+"/distanceNNData.json","w+") as f:
-            json.dump(finalData, f)
-
-def FunctionalModel():
-    X, y = loadANNData_3()
-    #print(y)
-    y = np.array(y, dtype="float64")
-    print(y)
-    i1 = np.array(X,dtype="float64")
-    print(i1[0])
-    input()
-    #i2 = np.array(X[1], dtype="float64")
-    #print(i2[0])
-    #input()
-    #i3 = np.array(X[2], dtype=object)
-    #print(i3[0])
-    #input()
-
-    print(i1[0].shape)
-    print(i1.shape)
-    input()
-    print(i1)
-    print("\n")
-    input()
-    '''
-    for i in range(0,600):
-        print(i1[i].shape,i2[i].shape,i3[i].shape)
-    input()
-    '''
-    #print(X[0])
-    #input()
-    #input1 = Input(shape=(16,))
-    '''
-    input2 = Input(shape=(17,))
-    #input0 = Concatenate()([input1, input2])
-    x = Dense(2)(input2)
-    hidden1 = Dense(10,activation='relu')(x)
-    hidden2 = Dense(8,activation='relu')(hidden1)
-    hidden3 = Dense(6, activation='relu')(hidden2)
-    output = Dense(1,activation='relu')(hidden3)
-    model = Model(inputs=input2, outputs=output)
-    model.summary()
-    model.compile(
-        loss="mean_squared_error",
-        optimizer="adam",
-        metrics=["mean_absolute_error"]
-    )
-    '''
-    model = Sequential()
-    model.add(Dense(1, input_dim=1, activation="relu"))
-    model.add(Dense(1, activation="linear"))
-    opt = SGD(learning_rate=0.01,momentum=0.9)
-    model.compile(
-        loss="mean_squared_logarithmic_error",
-        optimizer=opt,
-        metrics=["mse"]
-
-    )
-    x_train, x_test, y_train, y_test = train_test_split(i1, y, train_size=0.8, random_state=101)
-    print(x_train)
-    print(y_train)
-    input()
-    #print(x1_train[0],x2_train[0], y_train[0])
-    #input()
-    history = model.fit(x_train, y_train, batch_size=64, epochs=100, validation_split=0.2)
-    test_scores = model.evaluate(x_test, y_test, verbose=2)
-    print("Test loss:", test_scores[0])
-    print("Test accuracy:", test_scores[1])
-    print(test_scores)
-    predicted = model.predict(x_test)
-
-    print(predicted)
-    print(y_test)
-    for i in range(0,len(y_test)):
-        print(predicted[i], "and",y_test[i], "and", x_test[i])
-        input()
-
-def error_calculation(arr, identified):
-    '''
-        This function is similar to the check_dissonance function, but is different from it to an extent.
-        check_dissonance can be deleted, once this one is working properly.
-        Jesus it's such a mindfuck to explain what this is doing, but I will try.
-        So, this function in simple terms going to calculate the error likliness for each signal in a batch.
-        Obviously if there are more 'faulty' signals, then the other 'good' signal values are going to increase
-        as well. We don't want that. So, this function is going to be called recursively after the highest
-        error is going to be removed, and perhaps changed into a better representation of the signal. OR
-        this function could be used just to identify the best signals for trilateration.
-
-        The way it calcualtes the error is by calculating the otherNode locations to the node that the current signal
-        is received by. By the assumption that the lower the signal, the higher the 'drone' is and that means the otherNodes are
-        also higher, meaning it should have the highest nodeDistanceMean whatsoever.
-        Then we would check how much percentage is it needed to get to below and above the certain value, based on whether the signal
-        is lower or higher than the current one.
-        And then punishment / reward numerator is added, then it is all divided by a size (the highest probability it can get)-
-        Thus, giving us a cool value for error. But as said before, if there is a wrong signal in the batch, that is going to increase the
-        other signals' errors as well.
-    '''
-    #print(arr)
-    currSig = arr[0]
-    currDistS = arr[1]
-    currOtherSig = arr[2]
-    currOtherDistS = arr[3]
-    size = (len(currOtherSig))
-    numerator = 0
-    endPercent = 50
-    errorDistances = []
-    for i in range(0,len(currOtherSig)):
-        compare = currOtherDistS[i]
-        val0 = currDistS * 1.0
-        val1 = currDistS
-        val2 = currDistS
-        percent = 0
-        percent2 = 0
-        print("\n")
-        print(i,". Run")
-        print("#################################")
-
-        for j in range(0,endPercent):
-            if(val0 *(1+(j/100)) < compare):
-                val1 = val0 *(1+((j+1)/100))
-                percent = j+1
-            if(val0 *(1-(j/100)) > compare):
-                val2 = val0 *(1-((j+1)/100))
-                percent2 = j+1
-        if(currOtherSig[i] > currSig):
-
-            print("\n")
-            print(currOtherSig[i], "is bigger than ", currSig)
-            print(currOtherSig[i], currSig)
-            print(val0, compare)
-            print(val1, compare, percent)
-            print(val2, compare, percent2)
-            print("\n")
-
-
-            if(val0 < compare):
-                if(insideIdentified_2(identified, compare) == False):
-                    addNum = (0.1*(percent)/endPercent)
-                    numerator += addNum
-                    print("Possible dissonance")
-                    print(val0, " is less than ", compare, "so we add", addNum)
-                else:
-                    print("Possible dissonance, but ")
-                    print("Inside the identified error signals")
-                    addNum = (0.025*(endPercent-percent)/endPercent)
-                    numerator += addNum
-                    print(val0, "is less than", compare, "but", currOtherSig[i], "is in identified error signals")
-                    print("So we add", addNum)
-            else:
-                addNum = (0.05*(endPercent-percent2)/endPercent)
-                numerator += addNum
-                print("Good for first")
-                print(val0, " is bigger than ", compare, "so we add", addNum)
-
-        if(currOtherSig[i] < currSig):
-
-            print("\n")
-            print(currOtherSig[i], "is less than ", currSig)
-            print(currOtherSig[i], currSig)
-            print(val0, compare)
-            print(val1, compare, percent)
-            print(val2, compare, percent2)
-            print("\n")
-
-
-            if(val0 > compare):
-                if(insideIdentified_2(identified, compare) == False):
-                    addNum = (0.1*(percent2)/endPercent)
-                    numerator += addNum
-                    print("Possible dissonance")
-                    print(val0, " is bigger than ", compare, "so we add", addNum)
-                else:
-                    print("Possible dissonance, but ")
-                    print("Inside the identified error signals")
-                    addNum = (0.025*(endPercent-percent2)/endPercent)
-                    numerator += addNum
-                    print(val0, "is bigger than", compare, "but", currOtherSig[i], "is in identified error signals")
-                    print("So we add", addNum)
-            else:
-                addNum = (0.05 *(endPercent-percent)/endPercent)
-                numerator += addNum
-
-                print("Good for first")
-                print(val0, " is less than ", compare, "so we add", addNum)
-
-        if(currOtherSig[i] == currSig):
-            print("\n")
-            print(currOtherSig[i], "is equal to ", currSig)
-            print(currOtherSig[i], currSig)
-            print(val0, compare)
-            print(val1, compare, percent)
-            print(val2, compare, percent2)
-            print("\n")
-
-            if(insideIdentified_2(identified, compare) == False):
-                if(val0 > compare):
-                    addNum = (0.1*(percent2)/endPercent)
-                    numerator += addNum
-                else:
-                    addNum = (0.1 *(percent)/endPercent)
-                    numerator += addNum
-            else:
-                if(val0 > compare):
-                    addNum = (0.05*(percent2)/endPercent)
-                    numerator += addNum
-                else:
-                    addNum = (0.05*(percent)/endPercent)
-                    numerator += addNum
-
-        #input()
-        print("#################################")
-        print("\n")
-
-        '''
-        if(addNum != 0):
-            dist1 = calculateDist_2(currSig)
-            dist2 = calculateDist_2(currOtherSig[i])
-            err = abs(dist1-dist2)
-            if(currOtherSig[i] < currSig):
-                err = err*(percent2/endPercent)
-            elif(currOtherSig[i] > currSig):
-                err = -err*(percent/endPercent)
-            else:
-                if(percent2 > percent):
-                    err = err*(percent2/endPercent)
-                else:
-                    err = -err*(percent/endPercent)
-            #errorDistances.append([addNum, err, currOtherSig[i])
-            if(len(errorDistances) == 0):
-                errorDistances.append([currSig, arr[len(arr)-2]])
-        '''
-    '''
-        The error distances need to compare each signal error % in a bacth, to successfully indentify
-        the signal(s) that are wrong. After identifying those signals, we can make a proper guess at the
-        error distances. Otherwise, it is going to just mess it up.
-    '''
-    '''
-        Another idea, is to recursively run this code, getting the highest error signal out of the batch,
-        calculating the percentage for error again (then the likely 'big' error node will not affect the others)
-        and then do it until we have only 3 signals left. Those 3 signals will be used for trilateration.
-    '''
-
-
-    #print(numerator)
-    #print(size)
-
-    #print(errorDistances)
-    #print(round(numerator/size*100,2), "%")
-
-    return [arr, round(numerator/size*100,2)]
-
-def removeHighestError(arr, arr2, identified):
-    arr2_sorted = sorted(arr2)
-    print("\n")
-    print("removeErrorFunction")
-    print("\n")
-    lowB = deviation(arr2, arr2_sorted[0])
-    uppB = deviation(arr2, arr2_sorted[len(arr2_sorted)-1])
-    if(len(identified) == 0):
-        if(lowB > uppB):
-            index = findIndex(arr2, arr2_sorted[0])
-            print(arr2[index])
-            print(arr[index])
-            identified.append([arr[index], index])
-        else:
-            index = findIndex(arr2, arr2_sorted[len(arr2)-1])
-            #print(arr[index])
-            #print(arr2[index])
-            identified.append([arr[index], index])
-        print("\n")
-        print(arr)
-        print(arr2)
-        print(identified)
-        print("\n")
-        return arr, identified
-    elif(len(arr) - len(identified) == 3):
-        print("\n")
-        print(arr)
-        print("\n")
-        print(identified)
-        for ident in identified:
-            arr.remove(ident[0])
-
-        return arr, identified
-    else:
-        inside = []
-        outside = []
-        lowest = 100
-        highest = -1
-        for err in range(0,len(arr2)):
-
-            if(insideIdentified(identified, err)):
-                continue
-            else:
-                if(arr2[err] > highest):
-                    highest = arr2[err]
-                    indexHigh = err
-                if(arr2[err] < lowest):
-                    lowest = arr2[err]
-                    indexLow = err
-        print(lowest, highest)
-        print(indexLow, indexHigh)
-        lowB = deviation(arr2, lowest)
-        uppB = deviation(arr2, highest)
-
-        #Finding which one to take in
-        print(arr2[indexLow])
-        print(arr[indexLow])
-        print(arr[indexHigh])
-        print(arr2[indexHigh])
-        print(lowB, uppB)
-        print(arr2_sorted)
-        if(lowB > uppB):
-            print(arr2[indexLow])
-            print(arr[indexLow])
-            identified.append([arr[indexLow], indexLow])
-        else:
-            print(arr[indexHigh])
-            print(arr2[indexHigh])
-            identified.append([arr[indexHigh], indexHigh])
-        print("\n")
-        print(arr)
-        print(arr2)
-        print(identified)
-        print("\n")
-        return arr, identified
-
-    #input()
-def deviation(arr, value):
-    total = 0
-    for i in range(0, len(arr)):
-        total += pow(arr[i]-value,2)
-    result = sqrt(total/(len(arr)-1))
-    return result
-
-def insideIdentified(arr, index):
-    for i in range(0, len(arr)):
-        if(arr[i][1] == index):
-            return True
-    return False
-
-def insideIdentified_2(arr, value):
-    for i in range(0, len(arr)):
-        if(arr[i][0][1] == value):
-            return True
-    return False
-
-def findIndex(arr, value):
-    for i in range(0, len(arr)):
-        if(arr[i] == value):
-            return i
-def testLowest(arr):
-    err = []
-    for i in range(0,len(arr)):
-        err.append(arr[i][6])
-    ident = []
-    while(len(arr) - len(ident) != 3):
-        highest = -1
-        highestIndex = None
-        for j in range(0,len(err)):
-            if(abs(err[j]) > highest and j not in ident):
-                highest = abs(err[j])
-                highestIndex = j
-        ident.append(highestIndex)
-    newImportant = []
-    for k in range(0,len(arr)):
-        if(k not in ident):
-            newImportant.append(arr[k])
-    return newImportant
-
-
-
-
-def loadANNData_3():
-    '''
-        Idea here would be gather the signal and then the calculated distance, then get what the error is
-        then train the signal and the error on the model. Model would try to predict the error that is
-        contributed to the signal~distance.
-    '''
-    grid, sections, nodes = loadSections()
-    month = "June"
-    pathName = "../Data/"+month+"/distanceNNData.json"
-    with open(pathName,"r") as f:
-        data = json.load(f)
-    iput1 = []
-    iput2 = []
-    iput3 = []
-    output = []
-    batchN = 0
-    a = 0
-    distanceSum = []
-    actualDist = []
-    allSignals = []
-    importantDs = []
-    goodCount = 0
-    badCount = 0
-    goodCount2 = 0
-    badCount2 = 0
-    dronePos = None
-    minimals = []
-    fileX = []
-    fileY = []
-    for index, i in enumerate(data["X"]):
-        actual = data['y'][index]
-        predicted = calculateDist_2(i[0])
-        err = actual - predicted
-        err = np.float64(err)
-        #print("SIGNAL VALUE TO BE MANUPULATED")
-        #print(i[0], i[1], i[2], actual, predicted, err, i[3])
-        #input()
-        #print(err)
-        importantD = [i[0], i[1], i[2], actual, predicted, err, i[3], i[4]]
-
-        importantD[1] = sum(importantD[1])/len(importantD[1])
-        if(dronePos != data["Z"][index]):
-            keptDronePos = dronePos
-            dronePos = data["Z"][index]
-        if(i[3] == batchN):
-            total = 0
-            for number in i[1]:
-                total += number
-            distanceSum.append([total/len(i[1]), a])
-            actualDist.append(actual)
-            allSignals.append(i[0])
-            importantDs.append(importantD)
-            a += 1
-
-
-        else:
-            for e in range(0,len(importantDs)):
-                addTotals = []
-                for j in range(0,len(distanceSum)):
-                    if(e == distanceSum[j][1]): continue
-                    addTotals.append(distanceSum[j][0])
-                importantDs[e].insert(3, addTotals)
-            #print(distanceSum, actualDist, allSignals)
-            # This is where we calculate the value between that will be between -1 to 1, that will tell us
-            # how big the error could be
-            # The next value then will tell us how likely the error is
-            # Those two values are going to be used by the machine along with the signal, to predict the error
-            # Then the error is going to be added to the calcualted distance, giving a better result for the RSSI ~ Distance
-            utmCord = utm.from_latlon(keptDronePos[0],keptDronePos[1])
-            inside = pointToSection(utmCord[0], utmCord[1], sections)
-            if(inside != -1):
-                print("\n")
-                print("This is the data before selection: ")
-                print(importantDs)
-                print("\n")
-                identified = []
-                if(len(importantDs) < 3 or len(importantDs[0][3])== 0):
-                    canCont = True
-                else:
-                    canCont = False
-                while canCont == False:
-                    #print("Started again, mate!")
-                    errorCalcs = []
-                    newImportants = []
-                    for important in importantDs:
-                        newImportant = error_calculation(important, identified)
-                        errorCalcs.append(newImportant[1])
-                        newImportants.append(newImportant[0])
-                    #print("\n")
-                    #print(newImportants)
-                    print(errorCalcs)
-                    #input()
-                    print("\n")
-                    if(len(newImportants) == 3):
-                        canCont = True
-                    else:
-
-                        newImportants, identified = removeHighestError(newImportants, errorCalcs, identified)
-                        #Remove the commenting to from the testLowest, to test what would be the value,
-                        #if we chose all the best values for trilateration, turns out the function is working, just
-                        #trilateriation is not that powerful haha.
-
-                        #newImportants = testLowest(newImportants)
-
-                        #print("\n")
-                        #print("This is new Importants! After removing the highest error!")
-                        #print(newImportants)
-                        #print(identified)
-                        #print("\n")
-                        #input()
-                        importantDs = newImportants
-                        #input()
-                        '''
-                        for toPop in maxIndices:
-                            print("This is ",toPop)
-                            popped = newImportants.pop(toPop)
-                            popped = popped[1]
-                        importantDs = newImportants
-                        '''
-                        #popped = newImportants.pop(maximumErrorIndex)
-                        #popped = popped[1]
-                        #print(popped)
-                        #importantDs = removeHighestError(newImportants, popped)
-                print("So this is going to be used for Trilateration: ")
-
-                '''
-                    Gathering the data for a trilateration json ahhahaha
-                '''
-                fileData = {}
-                for fileE in importantDs:
-                    fileData[fileE[len(fileE)-1]] = fileE[0]
-                #print(fileData)
-                outerFileData = {}
-                outerFileData["time"] = random.randint(0,1000000)
-                outerFileData["tag"] = random.randint(0,1000000)
-                outerFileData["data"] = fileData
-
-                fileX.append(outerFileData)
-                fileY.append(keptDronePos)
-                print(importantDs)
-                print("\n")
-                #input()
-            #Decide if the idea is good or not here, batch data gathered
-            maxi2 = 0
-            mini2 = 10000000
-            for d in range(0, len(actualDist)):
-                if(maxi2 < actualDist[d]):
-                    maxi2 = actualDist[d]
-                    maxi2Index = d
-                if(mini2 > actualDist[d]):
-                    mini2 = actualDist[d]
-                    mini2Index = d
-            #Playing around with the maximum a bit
-            maxi = distanceSum[maxi2Index][0] * 1.35
-            maxiIndex = maxi2Index
-            mini = distanceSum[mini2Index][0] * 0.65
-            miniIndex = mini2Index
-
-            for n in range(0,len(distanceSum)):
-                if(maxi < distanceSum[n][0]):
-                    maxi = distanceSum[n][0]
-                    maxiIndex = n
-                if(mini > distanceSum[n][0]):
-                    mini = distanceSum[n][0]
-                    miniIndex = n
-            if(maxi2Index == maxiIndex):
-                #print("OMG TRUE!")
-                goodCount += 1
-
-            else:
-                #print("You son of a bitch that is not true")
-                badCount += 1
-            if(mini2Index == miniIndex):
-                goodCount2 +=1
-            else:
-                badCount2 += 1
-            #Reset the batch data, start with new batch
-            a = 0
-            batchN = i[3]
-            distanceSum = []
-            actualDist = []
-            allSignals = []
-            importantDs = []
-            total = 0
-            for number in i[1]:
-                total += number
-            distanceSum.append([total/len(i[1]),a])
-            actualDist.append(actual)
-            allSignals.append(i[0])
-            importantDs.append(importantD)
-            a += 1
-
-        iput1.append(np.array(i[0]))
-        output.append(np.array(err))
-    print(goodCount, badCount)
-    print(goodCount2, badCount2)
-    input()
-    print("\n")
-    print("\n")
-    for i in minimals:
-        print(i)
-        input()
-
-    assert(len(fileX) == len(fileY))
-    finalData = {}
-    finalData['X']=fileX
-    finalData['y']=fileY
-
-    #Save the file in a json, that will be used for modelling / training
-    with open("../Data/"+month+"/trilatData.json","w+") as f:
-            json.dump(finalData, f)
-    return iput1, output
-
-
-def loadANNData_2():
-    '''
-        Contains the data in format that the machine's gonna learn on.
-    '''
-    month = "June"
-    pathName = "../Data/"+month+"/distanceNNData.json"
-    with open(pathName,"r") as f:
-        data = json.load(f)
-    iput1 = []
-    iput2 = []
-    iput3 = []
-    output = []
-    maxi = 0
-    for index, i in enumerate(data["X"]):
-        #temp0 = []
-        temp0_ind = 0
-        temp1_ind = 0
-        '''
-        while len(temp0) != 15:
-            if(temp0_ind == 0):
-                temp0.append(i[0])
-                temp0_ind += 1
-            else:
-                temp0.append(np.float64(0))
-        '''
-        if(i[0] < -102): continue
-        #print(i[0], "and the actual distance away is: ", data['y'][index])
-        #print("\n")
-        #print(i[1])
-        #print("\n")
-        #print(i[2])
-        #input()
-        while len(i[1]) != 16:
-            i[1].append(np.float64(0))
-        while len(i[2]) != 17:
-            if(temp1_ind == 0):
-                i[2].insert(0,i[0])
-                temp1_ind +=1
-            else:
-                i[2].append(np.float64(0))
-        i[1].insert(0,np.float64(i[0]))
-        iput1.append(np.array(i[0]))
-        iput2.append(np.array(i[1]))
-        iput3.append(np.array(i[2]))
-        output.append(np.array(data['y'][index]))
-    return iput1, output
-def deviation(signals, signal):
-    total = 0
-    for i in range(0,len(signals)):
-        total = total + pow((signals[i]-signal),2)
-    return math.sqrt(total/len(signals))
-
-def getSignals(arr):
-    keys = arr.keys()
-    returned = {}
-    for key in keys:
-        returned[key] = arr[key][2]
-    return returned
-
-
-def getNodedistances(arr):
-    nodes = loadNodes(rewriteUTM=True)
-    keys = arr.keys()
-    skipKeys = []
-    #Adds the NodeIDs that has signals out of the equation range -103<RSSI<-72
-    for keyS in keys:
-        if(keyS == "3288000000"): key="3288e6"
-        currentNode = arr[keyS]
-        if(arr[keyS][2][0] < -101):
-            skipKeys.append(keyS)
-        if(arr[keyS][2][0] > -73):
-            skipKeys.append(keyS)
-    #This loop gathers the data so that the loadANNData functions can use it
-    #Returns a list, in the form [[Nodes distances away from the current node],[position of the current node],[signal corresponding to the node]]
-    #Also returns the list that contains the taken out NodeIDs
-    for key in keys:
-        if(key == "3288000000"): key="3288e6"
-        if(key in skipKeys):continue
-        currentNode = arr[key]
-        nodeDistances=[]
-        currentNodePos = [nodes[key]['Latitude'],nodes[key]['Longitude']]
-        utmCurrentNodePos = utm.from_latlon(currentNodePos[0], currentNodePos[1])
-        utmCurrentNodePosArr = np.array([np.float64(utmCurrentNodePos[0]),np.float64(utmCurrentNodePos[1])])
-        for node in nodes:
-            if(node == key): continue
-            if(node not in keys): continue
-            if(node in skipKeys): continue
-            nodePos = [nodes[node]['Latitude'],nodes[node]['Longitude']]
-            utmNodePos = utm.from_latlon(nodePos[0], nodePos[1])
-            utmNodePosArr = np.array([np.float64(utmNodePos[0]),np.float64(utmNodePos[1])])
-            aDist = np.linalg.norm(utmNodePosArr-utmCurrentNodePosArr)
-            nodeDistances.append(aDist)
-        arr[key][0] = nodeDistances
-        arr[key][2] = arr[key][2][0]
-    return arr, skipKeys
-
-
-def LoadMLPData(month="June"):
-    #This data will be loading RSS values and coordinates, so that the machine can 'learn' on it
-    pathName = "../Data/"+month+"/newData.json"
-    with open(pathName,"r") as f:
-        data = json.load(f)
-    batchSignals = []
-    targetPos = []
-
-    for index, item in enumerate(data["X"]):
-
-        #Keys please
-        keys = item["data"].keys()
-        signals= []
-        newSignals = []
-        #Loop through each item in X data set
-        for key in keys:
-            dronePos = (data["y"][index][0],data["y"][index][1])
-            #print(item["data"][key][2])
-            #input()
-            if(dronePos == (0,0) or dronePos == (0.754225181062586, -12.295563892977972) or dronePos == (4.22356791831445, -68.85519277364834)): continue
-            signal = item["data"][key][2][0]
-            signals.append(signal)
-            convertedPos = utm.from_latlon(data["y"][index][0],data["y"][index][1])
-            dronePos = [np.float64(convertedPos[0]),np.float64(convertedPos[1])]
-        if len(signals) > 3:
-            for i in range(0,3):
-                maxi = max(signals)
-                newSignals.append(np.float64(maxi))
-                signals.remove(maxi)
-
-            #newSignals = np.array(newSignals)
-            #dronePos = np.array(dronePos)
-            batchSignals.append(newSignals)
-            targetPos.append(dronePos)
-    assert(len(targetPos) == len(batchSignals))
-    #finalData = {}
-    #finalData['X']=batchSignals
-    #finalData['Y']=targetPos
-    #with open("../Data/June/MLPData.json","w+") as f:
-    #    json.dump(finalData, f)
-    #batchSignals = np.array(batchSignals)
-    #targetPos = np.array(targetPos)
-    return batchSignals, targetPos
-
-
-def loadANNData(month="June"):
-    pathName = "../Data/"+month+"/newEquationData.json"
-    with open(pathName,"r") as f:
-        data = json.load(f)
-    values=[]
-    distances=[]
-    count = 0
-    for i in range(0,len(data["Y"])):
-        if(data["Y"][i] < -102):
-            count += 1
-            continue
-        values.append(np.float64([data["Y"][i]]))
-        distances.append(np.float64(data["X"][i]))
-
-    print("This much was taken out:", count)
-    X = np.array(values)
-    Y = np.array(distances)
-    #Y = Y.astype(int)
-    return X, Y
-
-
-
-
-
 def newEquationData(month="June"):
     #I believe it would be nice to have it like this
     #We can derive other equations from other months as well if needed
@@ -1630,7 +797,8 @@ def deriveEquation(month="June"):
 
     return([a,b,c])
 '''
-    Part for the algorithm that tries to optimize the signal strength values - currently not working properly
+    Part for the algorithm that tries to optimize the signal strength values - working to an extent - but I don't think this is the path forward.
+    We can keep this for negative results.
 
 '''
 
@@ -1753,6 +921,10 @@ def calculateDist_3(RSSI):
     model = pickle.load(open('anndistance.sav','rb'))
     y = model.predict([[np.float64(RSSI)]])
     return y[0]
+
+'''
+    Function that can be used to calculate the signal from distance. This will be relevant for the proximity kind of model.
+'''
 def calculateRSSI(distance):
     rssi = (np.exp((distance*-0.00568791789392432))*29.797940785785794)-102.48720932300988
     return rssi
@@ -1778,14 +950,4 @@ def calculateDist(RSSI):
 
 if __name__=="__main__":
     #associateJuneData(newData=True)
-    #deriveEquation()
-    #loadModelData()
-    #rewriteMarchData()
-    #print(multilat.predictions(rssiThreshold=-102,keepNodeIds=True, isTriLat = False, month="June"))
-    #associateJuneData(newData=True)
-    #LoadMLPData()
-    #loadANNData()
-    #LoadSeqData()
-    #loadANNData_2()
-    loadANNData_3()
-    #FunctionalModel()
+    print(multilat.predictions(rssiThreshold=-102,keepNodeIds=True, isTriLat = False, month="June"))
